Add react-native-config
