+++ android/settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'mtgx'
+include ':react-native-config'
+project(':react-native-config').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-config/android')
 
 include ':app'
