+++ package.json
@@ -21,16 +21,17 @@
   },
   "dependencies": {
     "react": "16.0.0-alpha.12",
-    "react-native": "0.47.1"
+    "react-native": "0.47.1",
+    "react-native-config": "0.6.0"
   },
   "devDependencies": {
     "babel-jest": "20.0.3",
     "babel-preset-react-native": "2.1.0",
-    "codeclimate-test-reporter": "^0.5.0",
+    "codeclimate-test-reporter": "0.5.0",
     "jest": "20.0.4",
     "react-test-renderer": "16.0.0-alpha.12",
-    "snazzy": "^7.0.0",
-    "standard": "^10.0.3"
+    "snazzy": "7.0.0",
+    "standard": "10.0.3"
   },
   "jest": {
     "preset": "react-native"
